requirements.txt
streamlit == 1.49.1
python-dotenv == 1.1.1
requests == 2.32.5
pymed == 0.8.9
sentence-transformers == 5.1.1
groq == 0.32.0
